+++ Kernel/Devices/GPU/Console/ContiguousFramebufferConsole.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <Kernel/Devices/GPU/Console/ContiguousFramebufferConsole.h>
-#include <Kernel/TTY/ConsoleManagement.h>
+#include <Kernel/Devices/TTY/ConsoleManagement.h>
 
 namespace Kernel::Graphics {
