+++ Kernel/Devices/GPU/Console/GenericFramebufferConsole.cpp
@@ -6,7 +6,7 @@
  */
 
 #include <Kernel/Devices/GPU/Console/GenericFramebufferConsole.h>
-#include <Kernel/TTY/ConsoleManagement.h>
+#include <Kernel/Devices/TTY/ConsoleManagement.h>
 
 namespace Kernel::Graphics {
