+++ Kernel/Devices/TTY/TTY.cpp
@@ -13,8 +13,8 @@
 #include <Kernel/API/ttydefaults.h>
 #include <Kernel/API/ttydefaultschars.h>
 #include <Kernel/Debug.h>
+#include <Kernel/Devices/TTY/TTY.h>
 #include <Kernel/Interrupts/InterruptDisabler.h>
-#include <Kernel/TTY/TTY.h>
 #include <Kernel/UnixTypes.h>
 
 namespace Kernel {
