+++ Kernel/FileSystem/DevPtsFS/FileSystem.cpp
@@ -6,9 +6,9 @@
  */
 
 #include <Kernel/Devices/DeviceManagement.h>
+#include <Kernel/Devices/TTY/SlavePTY.h>
 #include <Kernel/FileSystem/DevPtsFS/FileSystem.h>
 #include <Kernel/FileSystem/DevPtsFS/Inode.h>
-#include <Kernel/TTY/SlavePTY.h>
 #include <Kernel/Time/TimeManagement.h>
 
 namespace Kernel {
