+++ Kernel/FileSystem/DevPtsFS/Inode.h
@@ -7,9 +7,9 @@
 #pragma once
 
 #include <AK/Types.h>
+#include <Kernel/Devices/TTY/SlavePTY.h>
 #include <Kernel/FileSystem/DevPtsFS/FileSystem.h>
 #include <Kernel/FileSystem/Inode.h>
-#include <Kernel/TTY/SlavePTY.h>
 
 namespace Kernel {
