+++ Kernel/Syscalls/fork.cpp
@@ -6,9 +6,9 @@
  */
 
 #include <Kernel/Debug.h>
+#include <Kernel/Devices/TTY/TTY.h>
 #include <Kernel/FileSystem/Custody.h>
 #include <Kernel/Memory/Region.h>
-#include <Kernel/TTY/TTY.h>
 #include <Kernel/Tasks/PerformanceManager.h>
 #include <Kernel/Tasks/Process.h>
 #include <Kernel/Tasks/Scheduler.h>
